requirements.txt
pydantic>=2.8.0
cryptography>=43.0.0
pytest>=8.3.0
requests>=2.32.0
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
Pillow>=10.3.0
SQLAlchemy>=2.0.30
psycopg[binary]>=3.1.19
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.4.0
